split out into individual files and fix sofa bug
